+++ sources/core/objects/movable/Movable.h
@@ -29,7 +29,7 @@ public:
     virtual void BumpByGas(Dir dir, bool inside = false) override;
 public:
     Vector KV_SAVEABLE(force_);
-    int KV_SAVEABLE(force_progress_);
+    qint64 KV_SAVEABLE(force_progress_);
     bool KV_SAVEABLE(anchored_);
     int KV_SAVEABLE(last_move_);
     int KV_SAVEABLE(tick_speed_);
